refactored state init
