+++ cli/utils.ts
@@ -1,5 +1,4 @@
 import { resolve } from "../deps.ts";
-import { State } from "./state.ts";
 
 export const parsePath = (...path: string[]): string => {
   if (
@@ -11,29 +10,13 @@ export const parsePath = (...path: string[]): string => {
   return "file://" + resolve(...path);
 };
 
-export const queryHandler = async (
-  queryString: string,
-  state: State,
-  queryfn: (query: string) => any,
-) => {
-  const queries = queryString.trim().split(/(?<!\\);/);
-
-  if (queries[queries.length - 1]?.trim() === "") queries.pop();
-
-  state.debug(queries, "Queries");
-
-  const results = [];
-
-  for (let query of queries) {
-    query = query.trim().replace("\\;", ";");
-    state.debug(query, "Query");
-
-    const result = await queryfn(query + ";");
-
-    results.push(result);
+export const safeConfigImport = async (
+  file: string,
+): Promise<any | undefined> => {
+  try {
+    const configRaw = await import(file);
+    return configRaw.default;
+  } catch (e) {
+    return;
   }
-
-  state.debug(results, "Query result");
-
-  return results;
 };
